Added global early termination due to time and fixed a function call where to arguments were passed in reverse order. Makefile upgrade.
Fuction was to add a TT entry, time has been set to 0.8 seconds per
move. If a search is aborted, resulted move will not be used, but it
will be printed. make release or make all to produce optimized ouput for
both client/server parallel and serial, make parallel to rebuild
parallel agents (unoptimized) and make serial to rebuild serial agents
(unoptimized), make debug to rebuild parallel and serial unoptimized.